backend/requirements.txt
Flask==3.0.2
flask-cors
requests
google-cloud-vision